Fix preview image overlaps a call button during video call
If a device supports hiding the navigation bar and it is enabled,
preview image overlaps a call button during video call. Because the
preview layout doesn't care about the visibility of the navigation bar.

To fix the problem, use APIs of WindowInsets that care about the
visibility of the navigation bar.

Bug: 80376182
Test: manual - Checked the preview image doesn't overlap a call button
when the navigation bar is not visible.

Change-Id: I72c9d504f0485c24a6be09a50869a8248745623f
Signed-off-by: Sipun Ku Mahanta <sipunkumar85@gmail.com>