Fix issue that the hold label is not shown during remote video holding
Hold label is not shown even if a remote party holds a video call
because VideoCallFragment#updateRemoteOffView is not called. A condition
checking shouldShowRemote whether to call the method was added in

To solve this issue, checks isRemotelyHeld that indicates hold status of
remote party and calls #updateRemoteOffView if it is updated.

Test: manual - Check that the hold label is shown when a remote party
holds a video call
Bug: 110244504

Change-Id: If86f206efa392b1da9527bd14d481728279729c4
Signed-off-by: Sipun Ku Mahanta <sipunkumar85@gmail.com>